requirements.txt
apache-airflow==1.10.3
psycopg2==2.8.3